+++ home/vue_app/src/components/about-page/AboutPage.vue
@@ -1,0 +1,20 @@
+<template>
+    <div>
+        <SparcFooter/>
+    </div>
+</template>
+
+<script>
+    import SparcFooter from "../footer/Footer.vue";
+
+    export default {
+        name: 'about-page',
+        components: {
+            SparcFooter
+        }
+    }
+</script>
+
+<style>
+
+</style>
